GIF: Change background colour handling.
We used to memset a colour we looked up in the colour table when
clearing the previous frame.  This didn't make sense, because the
colour is four bytes wide.  Now we write the actual colour in
properly.

Also, the background color comes from the global colour table
(if one exists), rather than from the local colour table.  If
there is no colour table, black is used.